Split selection edit fade operations
