+++ src/app/controller/library/selection_edits/ops.rs
@@ -1,28 +1,12 @@
-use super::FadeDirection;
 use super::buffer::SelectionEditBuffer;
-use crate::selection::FadeParams;
 
-const MIN_MUTE_FADE_SECS: f32 = 0.002;
+mod fades;
+mod selection_fades;
 
-/// Inputs for applying gain and fade envelopes within a selected frame span.
-pub(crate) struct SelectionFadeRequest<'a> {
-    /// Audio samples to update in place.
-    pub samples: &'a mut [f32],
-    /// Channel count for interleaved sample layout.
-    pub channels: usize,
-    /// Sample rate used to translate minimum mute-fade durations into frames.
-    pub sample_rate: u32,
-    /// Inclusive selection start frame.
-    pub start_frame: usize,
-    /// Exclusive selection end frame.
-    pub end_frame: usize,
-    /// Gain applied uniformly across the selected span before edge fades.
-    pub selection_gain: f32,
-    /// Optional fade-in parameters for the left edge.
-    pub fade_in: Option<FadeParams>,
-    /// Optional fade-out parameters for the right edge.
-    pub fade_out: Option<FadeParams>,
-}
+#[cfg(test)]
+pub(crate) use fades::fade_factor;
+pub(crate) use fades::{apply_directional_fade, apply_edge_fades};
+pub(crate) use selection_fades::{SelectionFadeRequest, apply_selection_fades};
 
 pub(crate) fn crop_buffer(buffer: &mut SelectionEditBuffer) -> Result<(), String> {
     let cropped = slice_frames(
@@ -106,201 +90,7 @@ pub(crate) fn slice_frames(
     cropped
 }
 
-pub(crate) fn apply_directional_fade(
-    samples: &mut [f32],
-    channels: usize,
-    start_frame: usize,
-    end_frame: usize,
-    direction: FadeDirection,
-) {
-    let channels = channels.max(1);
-    let total_frames = samples.len() / channels;
-    let (clamped_start, clamped_end) = clamped_selection_span(total_frames, start_frame, end_frame);
-    if clamped_end <= clamped_start {
-        return;
-    }
-    apply_fade_ramp(
-        samples,
-        channels,
-        clamped_start,
-        clamped_end,
-        direction,
-        0.5,
-    );
-    match direction {
-        FadeDirection::LeftToRight => {
-            apply_muted_selection(samples, channels, clamped_end, total_frames);
-        }
-        FadeDirection::RightToLeft => {
-            apply_muted_selection(samples, channels, 0, clamped_start);
-        }
-    }
-}
-
-/// Apply fade-in and fade-out ramps at the edges of the selected span.
-pub(crate) fn apply_edge_fades(
-    samples: &mut [f32],
-    channels: usize,
-    start_frame: usize,
-    end_frame: usize,
-    fade_frames: usize,
-) {
-    let channels = channels.max(1);
-    let total_frames = samples.len() / channels;
-    let (clamped_start, clamped_end) = clamped_selection_span(total_frames, start_frame, end_frame);
-    if clamped_end <= clamped_start {
-        return;
-    }
-    let selection_frames = clamped_end - clamped_start;
-    let fade_frames = fade_frames.min(selection_frames / 2);
-    if fade_frames == 0 {
-        return;
-    }
-    let denom = (fade_frames.saturating_sub(1)).max(1) as f32;
-    // Use default curve of 0.5 for edge fades
-    let curve = 0.5;
-    for i in 0..fade_frames {
-        let t = i as f32 / denom;
-        let factor = fade_factor(fade_frames, t, FadeDirection::RightToLeft, curve);
-        let frame = clamped_start + i;
-        for ch in 0..channels {
-            let idx = frame * channels + ch;
-            if let Some(sample) = samples.get_mut(idx) {
-                *sample *= factor;
-            }
-        }
-    }
-    for i in 0..fade_frames {
-        let t = if fade_frames == 1 {
-            1.0
-        } else {
-            i as f32 / denom
-        };
-        let factor = fade_factor(fade_frames, t, FadeDirection::LeftToRight, curve);
-        let frame = clamped_end.saturating_sub(fade_frames) + i;
-        for ch in 0..channels {
-            let idx = frame * channels + ch;
-            if let Some(sample) = samples.get_mut(idx) {
-                *sample *= factor;
-            }
-        }
-    }
-}
-
-/// Apply optional fade-in and fade-out ramps within the selection bounds.
-/// A minimal fade is applied when an outer extension is present but the fade length is zero.
-pub(crate) fn apply_selection_fades(request: SelectionFadeRequest<'_>) {
-    let SelectionFadeRequest {
-        samples,
-        channels,
-        sample_rate,
-        start_frame,
-        end_frame,
-        selection_gain,
-        fade_in,
-        fade_out,
-    } = request;
-    let channels = channels.max(1);
-    let total_frames = samples.len() / channels;
-    let (clamped_start, clamped_end) = clamped_selection_span(total_frames, start_frame, end_frame);
-    if clamped_end <= clamped_start {
-        return;
-    }
-    let selection_frames = clamped_end - clamped_start;
-    let min_fade_frames = ((sample_rate.max(1) as f32) * MIN_MUTE_FADE_SECS)
-        .round()
-        .max(1.0) as usize;
-    if (selection_gain - 1.0).abs() > f32::EPSILON {
-        for frame in clamped_start..clamped_end {
-            let base = frame * channels;
-            for ch in 0..channels {
-                if let Some(sample) = samples.get_mut(base + ch) {
-                    *sample *= selection_gain;
-                }
-            }
-        }
-    }
-    if let Some(fade_in) = fade_in {
-        let base_fade_frames = ((selection_frames as f32) * fade_in.length)
-            .round()
-            .clamp(0.0, selection_frames as f32) as usize;
-        let fade_frames = if base_fade_frames > 0 {
-            base_fade_frames
-        } else if fade_in.mute > 0.0 {
-            min_fade_frames.min(selection_frames)
-        } else {
-            0
-        };
-        if fade_frames > 0 {
-            let mute_frames = ((selection_frames as f32) * fade_in.mute).round().max(0.0) as usize;
-            if mute_frames > 0 {
-                let mute_start = clamped_start.saturating_sub(mute_frames);
-                if mute_start < clamped_start {
-                    apply_fade_ramp(
-                        samples,
-                        channels,
-                        mute_start,
-                        clamped_start,
-                        FadeDirection::LeftToRight,
-                        fade_in.curve,
-                    );
-                }
-            }
-            let fade_end = clamped_start.saturating_add(fade_frames).min(clamped_end);
-            if fade_end > clamped_start {
-                apply_fade_ramp(
-                    samples,
-                    channels,
-                    clamped_start,
-                    fade_end,
-                    FadeDirection::RightToLeft,
-                    fade_in.curve,
-                );
-            }
-        }
-    }
-    if let Some(fade_out) = fade_out {
-        let base_fade_frames = ((selection_frames as f32) * fade_out.length)
-            .round()
-            .clamp(0.0, selection_frames as f32) as usize;
-        let fade_frames = if base_fade_frames > 0 {
-            base_fade_frames
-        } else if fade_out.mute > 0.0 {
-            min_fade_frames.min(selection_frames)
-        } else {
-            0
-        };
-        if fade_frames > 0 {
-            let mute_frames = ((selection_frames as f32) * fade_out.mute).round().max(0.0) as usize;
-            if mute_frames > 0 {
-                let mute_end = clamped_end.saturating_add(mute_frames).min(total_frames);
-                if clamped_end < mute_end {
-                    apply_fade_ramp(
-                        samples,
-                        channels,
-                        clamped_end,
-                        mute_end,
-                        FadeDirection::RightToLeft,
-                        fade_out.curve,
-                    );
-                }
-            }
-            let fade_start = clamped_end.saturating_sub(fade_frames).max(clamped_start);
-            if fade_start < clamped_end {
-                apply_fade_ramp(
-                    samples,
-                    channels,
-                    fade_start,
-                    clamped_end,
-                    FadeDirection::LeftToRight,
-                    fade_out.curve,
-                );
-            }
-        }
-    }
-}
-
-fn clamped_selection_span(
+pub(super) fn clamped_selection_span(
     total_frames: usize,
     start_frame: usize,
     end_frame: usize,
@@ -308,66 +98,6 @@ fn clamped_selection_span(
     let clamped_start = start_frame.min(total_frames);
     let clamped_end = end_frame.min(total_frames);
     (clamped_start, clamped_end)
-}
-
-fn apply_fade_ramp(
-    samples: &mut [f32],
-    channels: usize,
-    clamped_start: usize,
-    clamped_end: usize,
-    direction: FadeDirection,
-    curve: f32,
-) {
-    let frame_count = clamped_end - clamped_start;
-    let denom = (frame_count.saturating_sub(1)).max(1) as f32;
-    for i in 0..frame_count {
-        let progress = i as f32 / denom;
-        let factor = fade_factor(frame_count, progress, direction, curve);
-        let frame = clamped_start + i;
-        for ch in 0..channels {
-            let idx = frame * channels + ch;
-            if let Some(sample) = samples.get_mut(idx) {
-                *sample *= factor;
-            }
-        }
-    }
-}
-
-pub(crate) fn fade_factor(
-    frame_count: usize,
-    progress: f32,
-    direction: FadeDirection,
-    curve: f32,
-) -> f32 {
-    if frame_count == 1 {
-        return 0.0;
-    }
-    let smoothed = apply_s_curve(progress.clamp(0.0, 1.0), curve);
-    let factor = match direction {
-        FadeDirection::LeftToRight => 1.0 - smoothed,
-        FadeDirection::RightToLeft => smoothed,
-    };
-    factor.clamp(0.0, 1.0)
-}
-
-/// Apply S-curve interpolation with adjustable tension.
-/// t: 0.0-1.0 input
-/// curve: 0.0 = linear, 0.5 = medium S-curve, 1.0 = maximum S-curve
-fn apply_s_curve(t: f32, curve: f32) -> f32 {
-    if curve <= 0.0 {
-        // Linear
-        return t;
-    }
-
-    // Blend between linear and smootherstep based on curve value
-    let smootherstep = {
-        let t2 = t * t;
-        let t3 = t2 * t;
-        t3 * (t * (t * 6.0 - 15.0) + 10.0)
-    };
-
-    // Interpolate between linear and smootherstep
-    t * (1.0 - curve) + smootherstep * curve
 }
 
 pub(crate) fn apply_muted_selection(
